Allow delays smaller than 50ms by setting delaybeforesend to None
Pexpect added[1] a default delay of 50ms before send, by default, to work
around an issue with passwords being echod.  This makes it impossible to have
delays smaller than this 50ms, currently.  This commit allows setting
delaybeforesend to None, which is recommended by Pexpect to remove this
additional delay hardcoded in Pexpect's code.

[1] - https://github.com/pexpect/pexpect/blob/master/pexpect/pty_spawn.py#L136-L150